created 3 new pages that descend from the catalogue page and created the logic to display what we want from each section page and linked to those three pages from the catalogue page
