+++ src/App.js
@@ -2,10 +2,14 @@
 import './App.css';
 import { BrowserRouter, Routes, Route, Link } from 'react-router-dom'
 
+
 //importing components
 import MainPage from './components/MainPage';
 import ProductPage from './components/ProductPage';
 import Catalogue from './components/Catalogue';
+import EyeProducts from './components/EyeProducts';
+import LipProducts from './components/LipProducts';
+import FaceProducts from './components/FaceProducts';
 
 function App() {
 
@@ -21,6 +25,9 @@ function App() {
           <Route path="/*" element={<MainPage />} />
           <Route path="/:productID" element={<ProductPage />} />
           <Route path="/catalogue" element={<Catalogue />} />
+          <Route path="/catalogue/eyeproducts" element={<EyeProducts />}/>
+          <Route path="/catalogue/lipproducts" element={<LipProducts />}/>
+          <Route path="/catalogue/faceproducts" element={<FaceProducts />}/>
         </Routes>
       </div>
     </BrowserRouter>
